rename auth to upstream-authentication
