+++ services/src/modules/upstream-authentication/get-auth-headers.ts
