+++ services/src/modules/upstream-authentication/index.ts
@@ -1,2 +1,3 @@
 export { getAuthHeaders } from './get-auth-headers';
 export { ActiveDirectoryAuth } from './active-directory-auth';
+export { AuthenticationConfig } from './types';
